requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
chromadb>=0.4.0
openai>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
streamlit>=1.28.0
pydantic>=2.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
sentence-transformers>=2.2.0
requests>=2.31.0
